Вынес обработку в NDK в отдельный ImageProcessor

+++ app/src/main/java/com/kudashov/opencv_android/image_processor/ImageProcessor.kt
@@ -4,7 +4,7 @@ import android.graphics.Bitmap
 
 interface ImageProcessor {
 
-    suspend fun blur(bitmap: Bitmap, sigma: Double): Bitmap
+    suspend fun blur(bitmap: Bitmap, sigma: Int): Bitmap
 
     suspend fun meanShift(bitmap: Bitmap): Bitmap
 }
